+++ changelog.md
@@ -3,6 +3,9 @@
 
 > **Note:** The dates included here are the dates of the commits that were made to the repository. The actual release dates may be different.
 
+## 19 September 2024 (dev)
+* Update the CSIRO renderer to v0.40.1
+
 ## 18 September 2024 (dev)
 * Update the CSIRO renderer to v0.40.0
 * Enhance the POC $extract Questionnaire tester based on initial testing